+++ tests/pallas/tpu_sparsecore_pallas_test.py
@@ -1754,7 +1754,11 @@ class VectorSubcoreTest(PallasSCTest):
     )
     def f(x):
       @self.kernel(
-          out_shape=jax.ShapeDtypeStruct(x.shape, x.dtype, vma={"x"}),
+          out_shape=jax.ShapeDtypeStruct(
+              x.shape,
+              x.dtype,
+              manual_type=jax.sharding.ManualAxisType(varying={"x"}),
+          ),
           mesh=plsc.VectorSubcoreMesh(
               core_axis_name="core", subcore_axis_name="subcore", num_cores=1
           ),